refactor(gateway): split provider authority facade
