+++ wmo/runtime/gateway/sqlite/store.py
@@ -13,7 +13,6 @@ from datetime import UTC, datetime
 from pathlib import Path
 
 from wmo.common.core.artifacts import Sha256, sha256_json
-from wmo.common.models import ConnectionConfig
 from wmo.runtime.gateway.auth import (
     FingerprintPepperFile,
     GatewayAuthError,
@@ -33,14 +32,10 @@ from wmo.runtime.gateway.contracts import (
 from wmo.runtime.gateway.interfaces import GatewayClock
 from wmo.runtime.gateway.sqlite.migrations import connect_database, initialize_database
 from wmo.runtime.gateway.sqlite.provider_authority import (
-    ProviderConnectionAuthority,
     ProviderConnectionBinding,
-    active_provider_connections,
     bind_alias_provider_connections,
-    bound_provider_connections,
-    disable_provider_connection,
-    upsert_provider_connection,
 )
+from wmo.runtime.gateway.sqlite.provider_store import ProviderConnectionStoreMixin
 
 
 class GatewayStoreError(ValueError):
@@ -75,7 +70,7 @@ class SystemGatewayClock:
         return time.monotonic()
 
 
-class SQLiteGatewayStore:
+class SQLiteGatewayStore(ProviderConnectionStoreMixin):
     """SQLite implementation of gateway authority and management operations."""
 
     def __init__(
@@ -131,70 +126,6 @@ class SQLiteGatewayStore:
                 ) VALUES (?, ?, ?, ?, ?)
                 """,
                 (organization_id, slug, display_name, now, now),
-            )
-
-    def upsert_provider_connection(
-        self,
-        *,
-        organization_id: str,
-        connection_id: str,
-        revision_id: str,
-        config: ConnectionConfig,
-        replace: bool = False,
-    ) -> tuple[bool, ProviderConnectionAuthority]:
-        """Create or explicitly revise one secret-free serving connection."""
-        with self._transaction() as connection:
-            return upsert_provider_connection(
-                connection,
-                organization_id=organization_id,
-                connection_id=connection_id,
-                revision_id=revision_id,
-                config=config,
-                replace=replace,
-                now=utc_text(self._clock.now()),
-            )
-
-    def provider_connections(
-        self,
-        *,
-        organization_id: str,
-    ) -> tuple[ProviderConnectionAuthority, ...]:
-        """Return active provider connections from SQLite authority."""
-        with self._connect() as connection:
-            return active_provider_connections(
-                connection,
-                organization_id=organization_id,
-            )
-
-    def alias_provider_connections(
-        self,
-        *,
-        organization_id: str,
-        alias_id: str,
-        alias_revision_id: str,
-    ) -> tuple[ProviderConnectionAuthority, ...]:
-        """Return exact provider revisions frozen into one alias revision."""
-        with self._connect() as connection:
-            return bound_provider_connections(
-                connection,
-                organization_id=organization_id,
-                alias_id=alias_id,
-                alias_revision_id=alias_revision_id,
-            )
-
-    def disable_provider_connection(
-        self,
-        *,
-        organization_id: str,
-        connection_id: str,
-    ) -> bool:
-        """Disable an unreferenced serving provider connection."""
-        with self._transaction() as connection:
-            return disable_provider_connection(
-                connection,
-                organization_id=organization_id,
-                connection_id=connection_id,
-                now=utc_text(self._clock.now()),
             )
 
     def create_identity(
